细节优化。

+++ DotNetCore/System/Web/HttpSessionState.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Http;
 
 namespace System.Web
 {
-    public class HttpSessionState:ISession
+    public class HttpSessionState : ISession
     {
         Microsoft.AspNetCore.Http.HttpContext context
         {
@@ -17,25 +17,60 @@ namespace System.Web
 
         internal HttpSessionState()
         {
-           
+
         }
 
-        public string SessionID { get { return context.Session.Id; } }
+        public string SessionID
+        {
+            get
+            {
+                if (context.Session == null)
+                {
+                    return null;
+                }
+                return context.Session.Id;
+            }
+        }
 
-        public bool IsAvailable => context.Session.IsAvailable;
+        public bool IsAvailable
+        {
+            get
+            {
+                if (context.Session == null)
+                {
+                    return false;
+                }
+                return context.Session.IsAvailable;
+            }
+        }
 
-        public string Id => context.Session.Id;
+        public string Id => SessionID;
 
-        public IEnumerable<string> Keys => context.Session.Keys;
+        public IEnumerable<string> Keys
+        {
+            get
+            {
+                if (context.Session == null)
+                {
+                    return null;
+                }
+                return context.Session.Keys;
+            }
+        }
+
 
         public void Clear()
         {
+            if (context.Session == null)
+            {
+                return;
+            }
             context.Session.Clear();
         }
 
         public Task CommitAsync(CancellationToken cancellationToken = default(CancellationToken))
         {
-             return context.Session.CommitAsync(cancellationToken = default(CancellationToken));
+            return context.Session.CommitAsync(cancellationToken = default(CancellationToken));
         }
 
         public Task LoadAsync(CancellationToken cancellationToken = default(CancellationToken))
@@ -45,17 +80,30 @@ namespace System.Web
 
         public void Remove(string key)
         {
+            if (context.Session == null)
+            {
+                return;
+            }
             context.Session.Remove(key);
         }
 
         public void Set(string key, byte[] value)
         {
+            if (context.Session == null)
+            {
+                return;
+            }
             context.Session.Set(key, value);
         }
 
         public bool TryGetValue(string key, out byte[] value)
         {
-           return context.Session.TryGetValue(key, out value);
+            if (context.Session == null)
+            {
+                value = null;
+                return false;
+            }
+            return context.Session.TryGetValue(key, out value);
         }
     }
 }
